Bug 2039903 - Use proper types for PBrowser::SynthesizeNativeMouseEvent. r=masayuki
Differential Revision: https://phabricator.services.mozilla.com/D300803

UltraBlame original commit: 63498adb03393f5e6782acf3d959655039b64cf1

+++ widget/PuppetWidget.cpp
@@ -344,9 +344,7 @@ nsresult PuppetWidget::SynthesizeNativeMouseEvent(
     return NS_ERROR_FAILURE;
   }
   mBrowserChild->SendSynthesizeNativeMouseEvent(
-      aPoint, static_cast<uint32_t>(aNativeMessage),
-      static_cast<int16_t>(aButton), static_cast<uint32_t>(aModifierFlags),
-      notifier.SaveCallback());
+      aPoint, aNativeMessage, aButton, aModifierFlags, notifier.SaveCallback());
   return NS_OK;
 }
